treba zmert ampak načeloma zgleda kot da dela

+++ interface.h
@@ -1,8 +1,8 @@
 #ifndef _INTERFACE_H
 #define _INTERFACE_H
 
-int read_input_up(int wanted_temp, float incremantor, float limit);
-int read_input_down(int wanted_temp, float incremantor, float limit);
+float read_input_up(float wanted_temp, float incremantor, float limit);
+float read_input_down(float wanted_temp, float incremantor, float limit);
 void set_display(float temp,int mode);
